Flying Etiquette Survey Data

Data from the results of a SurveyMonkey survey commissioned by FiveThirtyEight for the story 41 Percent of Fliers Say It’s Rude To Recline Your Airplane Seat.

fly

Format

A data frame with 1040 rows and 27 variables:

id

Respondent ID

flight_freq

How often do you travel by plane?

do_you_recline

Do you ever recline your seat when you fly?

height

How tall are you?

has_child_under_18

Do you have any children under 18?

three_seats_two_arms

n a row of three seats, who should get to use the two arm rests?

two_seats_one_arm

In a row of two seats, who should get to use the middle arm rest?

window_shade

Who should have control over the window shade?

rude_to_move_to_unsold_seat

Is it rude to move to an unsold seat on a plane?

rude_to_talk_to_neighbor

Generally speaking, is it rude to say more than a few words to the stranger sitting next to you on a plane?

six_hr_flight_leave_seat

On a six hour flight from NYC to LA, how many times is it acceptable to get up if you're not in an aisle seat?

reclining_obligation_to_behind

Under normal circumstances, does a person who reclines their seat during a flight have any obligation to the person sitting behind them?

rude_to_recline

Is it rude to recline your seat on a plane?

eliminate_reclining

Given the opportunity, would you eliminate the possibility of reclining seats on planes entirely?

rude_to_switch_seats_friends

Is it rude to ask someone to switch seats with you in order to be closer to friends?

rude_to_switch_seats_family

Is it rude to ask someone to switch seats with you in order to be closer to family?

rude_to_wake_neighbor_bathroom

Is it rude to wake a passenger up if you are trying to go to the bathroom?

rude_to_wake_neighbor_walk

Is it rude to wake a passenger up if you are trying to walk around?

rude_to_bring_baby

In general, is it rude to bring a baby on a plane?

rude_to_bring_unruly_child

In general, is it rude to knowingly bring unruly children on a plane?

use_electronics_takeoff

Have you ever used personal electronics during take off or landing in violation of a flight attendant's direction?

smoked_inflight

Have you ever smoked a cigarette in an airplane bathroom when it was against the rules?

gender

Gender

age

Age

household_income

Household Income

education

Education

region

Region

Source

https://github.com/fivethirtyeight/data/tree/master/flying-etiquette-survey